services/order-gateway/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
prometheus-fastapi-instrumentator>=6.1.0
httpx>=0.27.0
